FIX: Clean up config, move decorators to config module
